fix: moved the new tests to ScriptBuilderTest file
