+++ src/test/java/co/rsk/bitcoinj/script/ScriptTest.java
@@ -39,7 +39,6 @@ import java.io.InputStreamReader;
 import java.math.BigInteger;
 import java.nio.charset.StandardCharsets;
 import java.util.*;
-import java.util.stream.Collectors;
 
 import org.hamcrest.core.IsNot;
 import org.junit.Assert;
@@ -80,67 +79,6 @@ public class ScriptTest {
         byte[] deriveAddressInBytes = Utils.sha256hash160(pubKeyInBytes);
         Address addressFromP2pkhScriptSig = new Address(MAINNET_PARAMS, deriveAddressInBytes);
         assertEquals(expectedAddress, addressFromP2pkhScriptSig.toString());
-    }
-
-    @Test
-    public void createMultiSigOutputScript_withTwentyPubKeys_shouldReturnAValidScript() {
-        // Arrange
-        int numberOfKeys = 20;
-
-        // Act & Assert
-        assertGivenNumberOfKeysCreatesAValidMultiSigOutputScript(numberOfKeys);
-    }
-
-    @Test
-    public void createMultiSigOutputScript_withFifteenPubKeys_shouldReturnAValidScript() {
-        // Arrange
-        int numberOfKeys = 15;
-
-        // Act & Assert
-        assertGivenNumberOfKeysCreatesAValidMultiSigOutputScript(numberOfKeys);
-    }
-
-    @Test
-    public void createMultiSigOutputScript_withSixteenPubKeys_shouldReturnAValidScript() {
-        // Arrange
-        int numberOfKeys = 16;
-
-        // Act & Assert
-        assertGivenNumberOfKeysCreatesAValidMultiSigOutputScript(numberOfKeys);
-    }
-
-    private static void assertGivenNumberOfKeysCreatesAValidMultiSigOutputScript(int numberOfKeys) {
-        List<BtcECKey> ecKeys = RedeemScriptUtils.getNKeys(numberOfKeys);
-        int threshold = numberOfKeys / 2 + 1;
-
-        Script multiSigOutputScript = ScriptBuilder.createMultiSigOutputScript(threshold, ecKeys);
-
-        // threshold (1) + pubkeys (numberOfKeys) + num of pubKeys (1) + OP_CHECKMULTISIG (1)
-        int expectedNumberOfChunks = getExpectedNumberOfChunks(numberOfKeys);
-        List<ScriptChunk> chunks = multiSigOutputScript.getChunks();
-        assertEquals(expectedNumberOfChunks, chunks.size());
-
-        int index = 0;
-        int actualThreshold = chunks.get(index++).decodeN();
-        assertEquals(threshold, actualThreshold);
-
-        List<byte[]> pubKeys = ecKeys.stream().map(BtcECKey::getPubKey).collect(Collectors.toList());
-        for (int i = 0; i < pubKeys.size(); i++) {
-            byte[] actualPubKeyInTheScript = chunks.get(i + 1).data;
-            assertArrayEquals(pubKeys.get(i), actualPubKeyInTheScript);
-            index++;
-        }
-
-        int actualTotalKeysNumber = chunks.get(index++).decodeN();
-        assertEquals(pubKeys.size(), actualTotalKeysNumber);
-
-        int actualMultiSigOpCode = chunks.get(index).opcode;
-        assertEquals(ScriptOpCodes.OP_CHECKMULTISIG, actualMultiSigOpCode);
-    }
-
-    private static int getExpectedNumberOfChunks(int pubKeys) {
-        // threshold (1) + pubkeys (N) + num of pubKeys (1) + OP_CHECKMULTISIG (1)
-        return pubKeys + 3;
     }
 
     @Test
